proprietary: Pin prebuilt odex/vdex files for stock camera
For an unknown reason, the stock camera needs to be speed compiled upon a fresh install or when cache is cleared, otherwise it crashes (SIGSEGV).

This usually happends in Android overnight with the device plugged in, but that's a bad UX.

Another option is to run 'cmd package compile -f -m speed com.sonyericsson.android.camera', but SELinux will not like it if we put in in a script.

Since dexpreopt'ing is currently not working for prebuilt .apk's (in priv-app at least), pin the prebuilt .odex and .vdex for the stock camera app instead.